Moved spells to their own files
