+++ src/main/java/com/johnwesthoff/bending/logic/AppletInputListener.java
@@ -14,12 +14,12 @@ import java.awt.event.MouseListener;
 import java.awt.event.MouseMotionListener;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import com.johnwesthoff.bending.spells.*;
 
 import javax.swing.JFrame;
 
 import com.johnwesthoff.bending.Client;
 import com.johnwesthoff.bending.Constants;
+import com.johnwesthoff.bending.spells.fire.Firebending;
 
 /**
  *
@@ -47,7 +47,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
             return;
         }
         int button = e.getButton();
-        double scale = pointer.owner.getHeight() / (double)Constants.HEIGHT_INT;
+        double scale = pointer.owner.getHeight() / (double) Constants.HEIGHT_INT;
         if (pointer.world != null) {
             pointer.world.mouseX = (int) ((e.getX()) / scale);
             pointer.world.mouseY = (int) (e.getY() / scale);
@@ -80,7 +80,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
 
     @Override
     public void mouseDragged(MouseEvent e) {
-        double scale = pointer.owner.getHeight() / (double)Constants.HEIGHT_INT;
+        double scale = pointer.owner.getHeight() / (double) Constants.HEIGHT_INT;
         if (pointer.world != null) {
             pointer.world.mouseX = (int) ((e.getX() + getStuff()) / scale);
             pointer.world.mouseY = (int) (e.getY() / scale);
@@ -89,7 +89,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
 
     @Override
     public void mouseMoved(MouseEvent e) {
-        double scale = pointer.owner.getHeight() / (double)Constants.HEIGHT_INT;
+        double scale = pointer.owner.getHeight() / (double) Constants.HEIGHT_INT;
         if (pointer.world != null) {
             pointer.world.mouseX = (int) ((e.getX() + getStuff()) / scale);
             pointer.world.mouseY = (int) (e.getY() / scale);
@@ -105,7 +105,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
             if (pointer.world == null) {
                 return;
             }
-            double scale = pointer.owner.getHeight() / (double)Constants.HEIGHT_INT;
+            double scale = pointer.owner.getHeight() / (double) Constants.HEIGHT_INT;
             pointer.world.mouseX = (int) ((scale * pointer.getWidth())
                     * ((e.getPoint().x - (pointer.getWidth() - pointer.getHeight())) / (double) pointer.getWidth()));
             // world.mouseX = ((e.getX()-(this.getWidth()-getHeight())/2)/scale);
@@ -148,7 +148,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
                                                 : pointer.world.mouseX,
                                         pointer.world.mouseY);
                                 if ((pointer.passiveList[pointer.spellBook].getName().equals("Fire Charge"))
-                                        && (pointer.spellList[pointer.spellBook][pointer.leftClick] instanceof Spell.Firebending)) {
+                                        && (pointer.spellList[pointer.spellBook][pointer.leftClick] instanceof Firebending)) {
                                     if (pointer.random.nextInt(5 - doublecast) == 0) {
                                         pointer.spellList[pointer.spellBook][pointer.leftClick].getAction(pointer);
                                     }
@@ -185,7 +185,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
                                                 : pointer.world.mouseX,
                                         pointer.world.mouseY);
                                 if ((pointer.passiveList[pointer.spellBook].getName().equals("Fire Charge"))
-                                        && (pointer.spellList[pointer.spellBook][pointer.leftClick] instanceof Spell.Firebending)) {
+                                        && (pointer.spellList[pointer.spellBook][pointer.leftClick] instanceof Firebending)) {
                                     if (pointer.random.nextInt(5 - doublecast) == 0) {
                                         pointer.spellList[pointer.spellBook][pointer.midClick].getAction(pointer);
                                     }
@@ -216,7 +216,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
                                                 : pointer.world.mouseX,
                                         pointer.world.mouseY);
                                 if ((pointer.passiveList[pointer.spellBook].getName().equals("Fire Charge"))
-                                        && (pointer.spellList[pointer.spellBook][pointer.rightClick] instanceof Spell.Firebending)) {
+                                        && (pointer.spellList[pointer.spellBook][pointer.rightClick] instanceof Firebending)) {
                                     if (pointer.random.nextInt(5 - doublecast) == 0) {
                                         pointer.spellList[pointer.spellBook][pointer.rightClick].getAction(pointer);
                                     }
@@ -333,7 +333,7 @@ public class AppletInputListener implements MouseListener, KeyListener, MouseMot
     public void keyReleased(KeyEvent e) {
         if (e.getKeyCode() == KeyEvent.VK_ENTER) {
             if (pointer.chatActive) {
-                pointer.chatMessage = pointer.username + ": " + pointer.chatMessage;
+                pointer.chatMessage = Client.username + ": " + pointer.chatMessage;
                 pointer.sendMessage(pointer.chatMessage);
                 if (pointer.chatMessage.contains("/suicide")) {
                     pointer.HP = 0;
