+++ src/main/java/com/johnwesthoff/bending/spells/lightning/LightningRod.java
@@ -1,0 +1,63 @@
+
+package com.johnwesthoff.bending.spells.lightning;
+
+import static com.johnwesthoff.bending.util.network.ResourceLoader.loadIcon;
+
+import java.nio.ByteBuffer;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import com.johnwesthoff.bending.Client;
+import com.johnwesthoff.bending.Server;
+import com.johnwesthoff.bending.logic.World;
+import com.johnwesthoff.bending.spells.Spell;
+
+public class LightningRod extends Lightning {
+    public LightningRod() {
+        ID = Server.LIGHTNING;
+        subID = 5;
+        locked = true;
+        unlockXP = 1250;
+        try {
+            icon = loadIcon("https://west-it.webs.com/spells/shocktower.png");
+        } catch (Exception ex) {
+            Logger.getLogger(Spell.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    @Override
+    public void getAction(Client app) {
+
+        X = app.world.x;
+        Y = app.world.y - World.head;
+        mx = 0;
+        my = 0;
+        maker = ID;
+        getMessage(app.out);
+    }
+
+    @Override
+    public int getCost() {
+        return 200;
+    }
+
+    @Override
+    public String getName() {
+        return "Lightning Rod";
+    }
+
+    @Override
+    public void getPassiveAction(Client app) {
+    }
+
+    @Override
+    public String getTip() {
+        return "<html>An intermediate lightning spell<br>Low-Moderate Energy Cost<br>Strongly affected by gravity<br>Drains enemy energy</html>";
+    }
+
+    @Override
+    public void getActionNetwork(World world, int px, int py, int mx, int my, int pid, int eid, ByteBuffer buf) {
+        throw new UnsupportedOperationException("Not supported yet.");
+    }
+}
+
