+++ src/main/java/com/johnwesthoff/bending/spells/water/WaterSpout.java
@@ -1,0 +1,64 @@
+
+package com.johnwesthoff.bending.spells.water;
+
+import static com.johnwesthoff.bending.util.network.ResourceLoader.loadIcon;
+
+import java.nio.ByteBuffer;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import com.johnwesthoff.bending.Client;
+import com.johnwesthoff.bending.Server;
+import com.johnwesthoff.bending.logic.World;
+import com.johnwesthoff.bending.spells.Spell;
+
+public class WaterSpout extends Waterbending {
+    public WaterSpout() {
+        ID = Server.WATERBENDING;
+        subID = 2;
+        try {
+            icon = (loadIcon("https://west-it.webs.com/spells/waterspout.png"));
+        } catch (Exception ex) {
+            Logger.getLogger(Spell.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    @Override
+    public void getAction(Client app) {
+        // throw new UnsupportedOperationException("Not supported yet.");
+        mx = app.world.mouseX + app.world.viewX;
+        my = app.world.mouseY + app.world.viewY;
+        X = mx;
+        Y = my;
+        if (app.world.ground.cellData[(int) X][(int) Y] != World.WATER) {
+            app.energico += this.getCost();
+            return;
+        }
+        if (app.world.isLiquid(app.world.x, app.world.y)) {
+            app.energico += 150;
+        }
+        maker = ID;
+        getMessage(app.out);
+    }
+
+    @Override
+    public int getCost() {
+        return 450;
+    }
+
+    @Override
+    public String getName() {
+        return "WaterSpout";
+    }
+
+    @Override
+    public String getTip() {
+        return "<html>An intermediate water spell<br>Moderate-High Energy Cost<br>Summon a waterspout from a water source</html>";
+    }
+
+    @Override
+    public void getActionNetwork(World world, int px, int py, int mx, int my, int pid, int eid, ByteBuffer buf) {
+        throw new UnsupportedOperationException("Not supported yet.");
+    }
+}
+
