+++ src/main/java/com/johnwesthoff/bending/spells/water/WaterbendingShard.java
@@ -1,0 +1,70 @@
+
+package com.johnwesthoff.bending.spells.water;
+
+import static com.johnwesthoff.bending.util.network.ResourceLoader.loadIcon;
+
+import java.nio.ByteBuffer;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import com.johnwesthoff.bending.Client;
+import com.johnwesthoff.bending.Server;
+import com.johnwesthoff.bending.logic.World;
+import com.johnwesthoff.bending.spells.Spell;
+
+public class WaterbendingShard extends Waterbending {
+    public WaterbendingShard() {
+        ID = Server.WATERBENDING;
+        subID = 4;
+        try {
+            icon = (loadIcon("https://west-it.webs.com/spells/watershard.png"));
+        } catch (Exception ex) {
+            Logger.getLogger(Spell.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    @Override
+    public void getAction(Client app) {
+        // throw new UnsupportedOperationException("Not supported yet.");
+        X = app.world.x;
+        Y = app.world.y - World.head;
+        if (app.world.isLiquid(app.world.x, app.world.y)) {
+            app.energico += 50;
+        }
+        mx = app.world.viewX;
+        my = app.world.mouseY - app.world.viewY;
+        double direction = 360 - Client.pointDir(app.world.x - app.world.viewX,
+                app.world.y - World.head - app.world.viewY, app.world.mouseX, app.world.mouseY);
+        // direction+=180;
+        mx = ((int) (Client.lengthdir_x(9, direction)));
+        my = ((int) (Client.lengthdir_y(9, direction)));
+        maker = ID;
+        getMessage(app.out);
+    }
+
+    @Override
+    public int getCost() {
+        return 250;
+    }
+
+    @Override
+    public String getName() {
+        return "Ice Shard";
+    }
+
+    @Override
+    public String getTip() {
+        return "<html>An intermediate water spell<br>Low-Moderate Energy Cost<br>Not affected by gravity<br>Deals light damage<br>Creates an ice block</html>";
+    }
+
+    @Override
+    public void getPassiveAction(Client app) {
+        // throw new UnsupportedOperationException("Not supported yet.");
+    }
+
+    @Override
+    public void getActionNetwork(World world, int px, int py, int mx, int my, int pid, int eid, ByteBuffer buf) {
+        throw new UnsupportedOperationException("Not supported yet.");
+    }
+}
+
